refactor stratigraphy
