+++ src/borehole_extraction/extraction/predictions/borehole_predictions.py
@@ -6,8 +6,8 @@ import io
 
 from borehole_extraction.extraction.groundwater.groundwater_extraction import GroundwatersInBorehole
 from borehole_extraction.extraction.metadata.metadata import BoreholeMetadata
-from borehole_extraction.extraction.stratigraphy.depths_materials_pairs.bounding_boxes import PageBoundingBoxes
 from borehole_extraction.extraction.stratigraphy.layer.layer import LayersInBorehole
+from borehole_extraction.extraction.stratigraphy.layer.page_bounding_boxes import PageBoundingBoxes
 
 
 @dataclasses.dataclass
