+++ src/borehole_extraction/extraction/stratigraphy/base_sidebar_entry/sidebar_entry.py
