+++ src/borehole_extraction/extraction/stratigraphy/layer/layer.py
@@ -2,8 +2,8 @@
 
 from dataclasses import dataclass
 
-from borehole_extraction.extraction.stratigraphy.depth.interval import DepthInterval
-from borehole_extraction.extraction.stratigraphy.depths_materials_pairs.bounding_boxes import PageBoundingBoxes
+from borehole_extraction.extraction.stratigraphy.interval.interval import DepthInterval
+from borehole_extraction.extraction.stratigraphy.layer.page_bounding_boxes import PageBoundingBoxes
 from borehole_extraction.extraction.util_extraction.data_extractor.data_extractor import (
     ExtractedFeature,
     FeatureOnPage,
