+++ src/borehole_extraction/extraction/stratigraphy/sidebar_classes/a_to_b_sidebar.py
@@ -5,11 +5,10 @@ from __future__ import annotations
 from dataclasses import dataclass
 
 import pymupdf
-from borehole_extraction.extraction.stratigraphy.depth.interval import AToBInterval
+from borehole_extraction.extraction.stratigraphy.interval.interval import AToBInterval, IntervalBlockGroup
 from borehole_extraction.extraction.util_extraction.geometry.geometry_dataclasses import Line
 from borehole_extraction.extraction.util_extraction.text.textline import TextLine
 
-from .interval_block_group import IntervalBlockGroup
 from .sidebar import Sidebar
