+++ src/borehole_extraction/extraction/stratigraphy/sidebar_extractor/depth_column_entry_extractor.py
@@ -2,8 +2,8 @@
 
 import re
 
-from borehole_extraction.extraction.stratigraphy.depth.a_to_b_interval_extractor import AToBIntervalExtractor
-from borehole_extraction.extraction.stratigraphy.sidebar.sidebarentry import DepthColumnEntry
+from borehole_extraction.extraction.stratigraphy.base_sidebar_entry.sidebar_entry import DepthColumnEntry
+from borehole_extraction.extraction.stratigraphy.interval.a_to_b_interval_extractor import AToBIntervalExtractor
 from borehole_extraction.extraction.util_extraction.text.textline import TextLine, TextWord
