+++ tests/test_layerdepths.py
@@ -1,8 +1,8 @@
 """Test suite for the LayerDepths class."""
 
 import pymupdf
+from borehole_extraction.extraction.stratigraphy.base_sidebar_entry.sidebar_entry import DepthColumnEntry
 from borehole_extraction.extraction.stratigraphy.layer.layer import DepthInterval
-from borehole_extraction.extraction.stratigraphy.sidebar.sidebarentry import DepthColumnEntry
 
 
 def test_line_anchor():  # noqa: D103
